chore: MainPage 코드 리팩토링 - 세부적으로 다 쪼개기
chore: #63
